requirements.txt
ultralytics>=8.4.0,<9
pyyaml>=6.0
pillow>=10.0
opencv-python-headless>=4.9
